@@ -14,11 +14,12 @@
         }
     ],
     "require": {
-        "php": ">=7.4",
-        "laravel/framework": "^6.20.26|^7.30.6|^8.0|^9.0|^10.0|^11.0"
+        "php": ">=8.1",
+        "laravel/framework": "^8.0|^9.0|^10.0|^11.0"
     },
     "require-dev": {
-        "orchestra/testbench": "^v4.0|^v5.0|^v6.0|^v7.0|^8.0|^9.0"
+        "orchestra/testbench": "^v6.0|^v7.0|^8.0|^9.0",
+        "pestphp/pest": "^2.36"
     },
     "autoload": {
         "psr-4": {
@@ -34,8 +35,13 @@
                 "MasterRO\\LaravelXSSFilter\\XSSFilterServiceProvider"
             ],
             "aliases": {
-                "XSSCleaner": "MasterRO\\LaravelXSSFilter\\XSSCleanerFacade"
+                "XSSCleaner": "MasterRO\\LaravelXSSFilter\\Facade\\XSSCleaner"
             }
+        }
+    },
+    "config": {
+        "allow-plugins": {
+            "pestphp/pest-plugin": true
         }
     }
 }
