@@ -5,31 +5,23 @@ declare(strict_types=1);
 namespace MasterRO\LaravelXSSFilter;
 
 use Illuminate\Support\ServiceProvider;
+use MasterRO\LaravelXSSFilter\Cleaner\Cleaner;
+use MasterRO\LaravelXSSFilter\Cleaner\CleanerConfig;
 
 class XSSFilterServiceProvider extends ServiceProvider
 {
-    /**
-     * Boot the service provider.
-     *
-     * @return void
-     */
-    public function boot()
+    public function boot(): void
     {
         $this->publishes([
-            __DIR__ . '/xss-filter.php' => config_path('xss-filter.php'),
+            __DIR__ . '/../config/xss-filter.php' => config_path('xss-filter.php'),
         ], 'config');
     }
 
-    /**
-     * Register the service provider.
-     *
-     * @return void
-     */
-    public function register()
+    public function register(): void
     {
-        $this->mergeConfigFrom(__DIR__ . '/xss-filter.php', 'xss-filter');
+        $this->mergeConfigFrom(__DIR__ . '/../config/xss-filter.php', 'xss-filter');
 
-        $this->app->singleton(Cleaner::class, static function () {
+        $this->app->scoped(Cleaner::class, static function () {
             return new Cleaner(CleanerConfig::fromArray(config('xss-filter')));
         });
     }
